Moving forms into own file, and cleaning comments on constants
